Refactor ScoreFunctionMap, Alternative, and WeightMap to use objective names instead of PrimitiveObjective objects.
This refactoring fixes a potential issue caused by the fact that Map keys are compared via memory location, rather than a deep compare in JavaScript. Almost everything in JavaScript is by reference, meaning that this is would not cause an issue almost all of the time. However, if someone were to explicitly copy a PrimitiveObjective, thus creating a new reference, then the copy would not be the same Map key as the original. Using strings fixes this potential problem.